+++ src/Database/Esqueleto/Experimental.hs
@@ -1166,6 +1166,9 @@ withRecursive baseCase unionKind recursiveCase = do
   Q $ W.tell mempty{sdCteClause = [clause]}
   pure refFrom
 
+{-# DEPRECATED ToAliasT "This type alias doesn't do anything. Please delete it. Will be removed in the next release. #-}
+type ToAliasT a = a
+
 -- Tedious tuple magic
 class ToAlias a where
   toAlias :: a -> SqlQuery a
@@ -1240,6 +1243,9 @@ instance ( ToAlias a
          ) => ToAlias (a,b,c,d,e,f,g,h) where
   toAlias x = to8 <$> (toAlias $ from8 x)
 
+
+{-# DEPRECATED ToAliasReferenceT "This type alias doesn't do anything. Please delete it. Will be removed in the next release. #-}
+type ToAliasReferenceT a = a
 
 -- more tedious tuple magic
 class ToAliasReference a where