+++ changelog.md
@@ -2,8 +2,7 @@
 =======
 - @arthurxavierx
   - [#221](https://github.com/bitemyapp/esqueleto/pull/221)
-    - Make `ToAliasT` an associated type family of the `ToAlias` type class
-    - Make `ToAliasReferenceT` an associated type family of the `ToAliasReference` type class
+    - Deprecate `ToAliasT` and `ToAliasReferenceT`
 
 3.4.0.0
 =======